HIR-59 refactor: review 도메인 rest controller 전환관련 테스트 수정 및 restdocs 테스트 추가
