+++ src/test/java/kr/binarybard/hireo/api/docs/ReviewControllerTest.java
@@ -1,7 +1,8 @@
-package kr.binarybard.hireo.web.review.controller;
+package kr.binarybard.hireo.api.docs;
 
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.*;
+import static org.springframework.restdocs.payload.PayloadDocumentation.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import org.junit.jupiter.api.BeforeEach;
@@ -10,11 +11,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.transaction.annotation.Transactional;
+import org.springframework.test.web.servlet.MockMvc;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import kr.binarybard.hireo.common.AcceptanceTest;
 import kr.binarybard.hireo.common.fixture.AccountFixture;
 import kr.binarybard.hireo.common.fixture.CompanyFixture;
 import kr.binarybard.hireo.common.fixture.ReviewFixture;
@@ -24,9 +24,11 @@ import kr.binarybard.hireo.web.company.domain.Company;
 import kr.binarybard.hireo.web.company.repository.CompanyRepository;
 import kr.binarybard.hireo.web.review.dto.ReviewRequest;
 
-@Transactional
 @WithMockUser(username = AccountFixture.TEST_EMAIL)
-class ReviewControllerTest extends AcceptanceTest {
+class ReviewControllerTest extends RestDocsConfiguration {
+
+	@Autowired
+	private MockMvc mockMvc;
 
 	@Autowired
 	private ObjectMapper objectMapper;
@@ -58,17 +60,12 @@ class ReviewControllerTest extends AcceptanceTest {
 				.content(objectMapper.writeValueAsString(reviewRequest)))
 			.andExpect(status().isCreated())
 			.andExpect(redirectedUrl("/companies/1/review/2701"))
-			.andDo(print());
-	}
-
-	@Test
-	@DisplayName("리뷰 내용이 10자 미만일 경우 리뷰 등록에 실패한다.")
-	void failToRegisterReviewWhenContentIsLessThan10() throws Exception {
-		ReviewRequest reviewRequest = ReviewFixture.createTestReviewRequestFail();
-		mockMvc.perform(post("/companies/1/review")
-				.contentType(MediaType.APPLICATION_JSON)
-				.content(objectMapper.writeValueAsString(reviewRequest)))
-			.andExpect(status().isBadRequest())
-			.andDo(print());
+			.andDo(document("review-register",
+				requestFields(
+					fieldWithPath("title").description("리뷰 제목"),
+					fieldWithPath("rating").description("리뷰 점수"),
+					fieldWithPath("content").description("리뷰 내용"),
+					fieldWithPath("postedAt").description("리뷰 게시 일자")
+				)));
 	}
 }
